requirements.txt
tornado==6.1
uvloop==0.51.2
toml==0.10.2
itsdangerous==2.0.0
ujson==1.35
argon2-cffi==20.1.0
asyncpg==0.23.0
